Add annotation table for each entity and add an API method to get InfosTransformateur by Transformateur

+++ src/main/java/com/istic/foodorigin/controller/TransformateurController.java
@@ -16,7 +16,7 @@ public class TransformateurController {
     private TransformateurService transformateurService;
 
     @GetMapping (path = "/{id}", produces = "application/json")
-    public Transformateur getTransformateurById (@PathVariable Integer id) {
+    public Transformateur getTransformateurById (@PathVariable Long id) {
         return transformateurService.getTransformateur(id);
     }
 }
